renamed css folder to styles
